apps/core_api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
pydantic==2.8.2
